+++ packages/terra-embedded-content-consumer/src/terra-dev-site/doc/example/CustomEventsConsumer.jsx
@@ -1,18 +1,20 @@
 import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
+// eslint-disable-next-line import/no-unresolved
 import EmbeddedContentConsumer from 'terra-embedded-content-consumer/lib/EmbeddedContentConsumer';
 
 const eventHandlers = [
   {
     key: 'EventA',
-    handler: () => {
+    handler: (consumer) => {
       document.getElementById('CustomEvents').style.border = 'thick dashed #0000FF';
+      consumer.trigger('Event-Reply', { eventReply: 'eventA', borderColor: '#0000FF' });
     },
   },
   {
     key: 'EventB',
-    handler: () => {
+    handler: (consumer) => {
       document.getElementById('CustomEvents').style.border = 'thick dashed #00FF00';
+      consumer.trigger('Event-Reply', { eventReply: 'eventB', borderColor: '#00FF00' });
     },
   },
 ];
@@ -20,7 +22,7 @@ const eventHandlers = [
 const CustomEventsConsumer = () => (
   <div id="CustomEvents">
     <EmbeddedContentConsumer
-      src="#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-events-provider"
+      src="#/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/custom-events-provider"
       options={{ iframeAttrs: { title: 'Custom events example' } }}
       eventHandlers={eventHandlers}
     />