+++ packages/terra-embedded-content-consumer/tests/wdio/embedded-content-consumer-spec.js
@@ -1,4 +1,3 @@
-/* global $ */
 const viewports = Terra.viewports('tiny', 'large');
 
 describe('Embedded Content Consumer', () => {
@@ -10,26 +9,11 @@ describe('Embedded Content Consumer', () => {
     Terra.should.matchScreenshot({ viewports });
   });
 
-  describe('fill', () => {
-    before(() => {
-      browser.url('#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/fill-consumer');
-    });
-
-    Terra.should.matchScreenshot({ viewports });
-  });
-
-  describe('on-ready', () => {
-    before(() => {
-      browser.url('#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/on-ready-consumer');
-    });
-
-    Terra.should.matchScreenshot({ viewports });
-  });
-
   describe('custom-event', () => {
     before(() => {
       browser.url('#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/custom-event-consumer');
       browser.waitForExist('#CustomEvent');
+      browser.pause(1000);
     });
 
     Terra.should.matchScreenshot({ viewports });
@@ -39,12 +23,17 @@ describe('Embedded Content Consumer', () => {
     before(() => {
       browser.url('#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/consumers/custom-events-consumer');
       browser.waitForExist('#CustomEvents');
-
-      // Waiting for events to execute.
-      browser.pause(5000);
+      browser.pause(1000);
     });
 
-    Terra.should.matchScreenshot({ viewports });
+    Terra.should.matchScreenshot('EventA', { viewports });
+
+    it('waits for Provider to trigger eventB', () => {
+      // Waiting for events to execute.
+      browser.pause(10000);
+    });
+
+    Terra.should.matchScreenshot('EventB', { viewports });
   });
 
   describe('data-status', () => {
@@ -54,9 +43,9 @@ describe('Embedded Content Consumer', () => {
 
     it('has mounted, launched, and authorized elements', () => {
       const timeout = browser.options.waitforTimeout + 2000;
-      browser.waitForExist('iframe[src="#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/providers/data-status-provider"]', timeout);
+      browser.waitForExist('iframe[src="#/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/data-status-provider"]', timeout);
 
-      const myFrame = $('iframe[src="#/raw/tests/terra-embedded-content-consumer/embedded-content-consumer/providers/data-status-provider"]').value;
+      const myFrame = browser.element('iframe[src="#/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/data-status-provider"]').value;
       browser.frame(myFrame);
 
       expect(browser.isExisting('#Mounted'));