+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -3,6 +3,18 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* EmbeddedContentProviderTestTemplate for testing embedded applicaitons
+
+### Changed
+* Updated existing provider examples to use the EmbeddedContentProviderTestTemplate
+* Changed Provider examples to use `.provider` extension to fix provider examples in terra-ui
+
+### Fixed
+* In the embedded-content-consumer event handler callback, pass the frame such that consumers can send custom events to the XFC provider of the embedded content
+
+### Removed
+* Dead code tests / examples that tested XFC implementation
 
 3.1.0 - (February 5, 2019)
 ------------------